requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==23.0
pycryptodome==3.16.0
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uWSGI==2.0.20